% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_dplyr.r
\name{s_filter}
\alias{s_filter}
\title{Modified version of dplyr's filter that uses string arguments}
\usage{
s_filter(.data, ...)
}
\arguments{
\item{.data}{the data frame (or similar object)}

\item{...}{string version of arguments to original dplyr function}
}
\description{
Modified version of dplyr's filter that uses string arguments
}

\examples{\dontrun{
  library(dplyrExtras)
  
  d = mtcars
  cols = c("mpg","cyl","hp:vs")
  s_select(d,cols)
  # also works and yields identical result...
  cols = c("mpg","cyl, hp:vs")
  s_select(d,cols)
  
  s_filter(d,"gear == 3","cyl == 8")
  s_arrange(d, "-mpg, gear, carb")
  gd = s_group_by(d,"cyl")
  s_summarise(gd, "mean(disp), max(disp)")
  s_mutate_if(d, "cyl==6, new.col=100")
  
}}
