% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_dplyr.r
\name{s_group_by}
\alias{s_group_by}
\title{Modified version of dplyr's group_by that uses string arguments}
\usage{
s_group_by(.data, ...)
}
\arguments{
\item{.data}{the data frame (or similar object)}

\item{...}{string version of arguments to original dplyr function}
}
\description{
Modified version of dplyr's group_by that uses string arguments
}

