% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{s_modify}
\alias{s_modify}
\title{Modified version of modify that uses string arguments}
\usage{
s_modify(.dt, ..., .envir = parent.frame())
}
\arguments{
\item{.dt}{the data.table / similar object that shall be modified}

\item{...}{string version of arguments of modify (see example)}

\item{.envir}{environment in which arguments will be evaluated (relevant if variables outside .dt are used)}
}
\description{
Modified version of modify that uses string arguments
}

