% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsummarise_each.r
\name{xsummarise_each_}
\alias{xsummarise_each_}
\title{extended version of summarise_each_ that allows alternative arrangement of outputs}
\usage{
xsummarise_each_(tbl, funs, vars, .long = "funs", .wide = "vars")
}
\description{
currently it is only checked whether "funs" is in .long or not.
}

