% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xsummarise_each}
\alias{xsummarise_each}
\title{extended version of summarise_each that allows alternative arrangement of outputs}
\usage{
xsummarise_each(tbl, funs, ..., .long = "funs", .wide = "vars")
}
\description{
currently it is only checked whether "funs" is in .long or not. if yes
then functions will be shown in different rows
}

\examples{\dontrun{
  library(dplyr)
  library(dplyrExtras)
  by_species <- group_by(iris,Species)
  summarise_each(by_species,funs(min, max), Petal.Width, Sepal.Width)
  xsummarise_each(by_species,funs(min, max), Petal.Width, Sepal.Width)
  xsummarise_each_q(by_species,funs(min, max), vars=c("Petal.Width", "Sepal.Width"))
}}
