% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xsummarise_each_}
\alias{xsummarise_each_}
\title{extended version of summarise_each_ that allows alternative arrangement of outputs}
\usage{
xsummarise_each_(tbl, funs, vars, .long = "funs", .wide = "vars")
}
\description{
currently it is only checked whether "funs" is in .long or not.
}

