% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xsummarise_each_q}
\alias{xsummarise_each_q}
\title{extended version of summarise_each_q that allows alternative arrangement of outputs}
\usage{
xsummarise_each_q(tbl, funs, vars, .long = "funs", .wide = "vars")
}
\description{
currently it is only checked whether "funs" is in .long or not.
}

